requirements.txt
streamlit==1.30.0
phidata==2.7.2
google-generativeai==0.8.3
tavily-python==0.5.0
python-dotenv==1.0.1
Pillow==10.2.0